dashCO2-web/requirements.txt
dash-bootstrap-components==0.11.2
dash-core-components==1.15.0
dash-daq==0.5.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
flask-sqlalchemy
numpy>=1.16.2
arrow<1
pyyaml
flask_admin
flask_httpauth